pyproject.toml
[project]
name = "commspt-scripts"
version = "0.1.0"
description = "LittleSkin CommSpt Scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2,socks]>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "tzdata>=2025.2",
]


[tool.ruff]
line-length = 120
